fix: fix package name
