+++ core/jpa-core/src/main/java/nettee/jpa/config/JpaConfig.java
@@ -1,4 +1,4 @@
-package me.nettee.config;
+package nettee.jpa.config;
 
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.Configuration;
