+++ core/jpa-core/src/main/java/nettee/jpa/support/LongBaseEntity.java
@@ -1,4 +1,4 @@
-package me.nettee.support;
+package nettee.jpa.support;
 
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -7,12 +7,11 @@ import jakarta.persistence.MappedSuperclass;
 import lombok.Getter;
 
 import java.io.Serializable;
-import java.util.UUID;
 
 @Getter
 @MappedSuperclass
-public abstract class UuidBaseEntity implements Serializable {
+public abstract class LongBaseEntity implements Serializable {
     @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    private UUID id;
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
 }
