+++ core/jpa-core/src/main/java/nettee/jpa/support/LongBaseTimeEntity.java
@@ -1,4 +1,4 @@
-package me.nettee.support;
+package nettee.jpa.support;
 
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
@@ -7,16 +7,15 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import java.io.Serializable;
 import java.time.Instant;
 
 @Getter
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener.class)
-public abstract class BaseTimeEntity implements Serializable {
+public abstract class LongBaseTimeEntity extends LongBaseEntity {
     @CreatedDate
     private Instant createdAt;
-    
+
     @LastModifiedDate
     private Instant updatedAt;
 }
