+++ core/jpa-core/src/main/java/nettee/jpa/support/UuidBaseEntity.java
@@ -1,4 +1,4 @@
-package me.nettee.support;
+package nettee.jpa.support;
 
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -7,11 +7,12 @@ import jakarta.persistence.MappedSuperclass;
 import lombok.Getter;
 
 import java.io.Serializable;
+import java.util.UUID;
 
 @Getter
 @MappedSuperclass
-public abstract class LongBaseEntity implements Serializable {
+public abstract class UuidBaseEntity implements Serializable {
     @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    @GeneratedValue(strategy = GenerationType.UUID)
+    private UUID id;
 }
