+++ services/board/api/exception/src/main/java/nettee/board/BoardCommandErrorCode.java
@@ -1,19 +1,22 @@
-package me.nettee;
+package nettee.board;
 
 import java.util.Map;
 import java.util.function.Supplier;
+
+import nettee.common.ErrorCode;
 import org.springframework.http.HttpStatus;
 
-public enum BoardQueryErrorCode implements ErrorCode {
+public enum BoardCommandErrorCode implements ErrorCode {
     BOARD_NOT_FOUND("게시물을 찾을 수 없습니다.", HttpStatus.NOT_FOUND),
     BOARD_GONE("더 이상 존재하지 않는 게시물입니다.", HttpStatus.GONE),
     BOARD_FORBIDDEN("권한이 없습니다.", HttpStatus.FORBIDDEN),
-    DEFAULT("게시물 조작 오류", HttpStatus.INTERNAL_SERVER_ERROR);
+    DEFAULT("게시물 조작 오류", HttpStatus.INTERNAL_SERVER_ERROR),
+    BOARD_ALREADY_EXIST("게시물이 이미 존재합니다.", HttpStatus.CONFLICT);
 
     private final String message;
     private final HttpStatus httpStatus;
 
-    BoardQueryErrorCode(String message, HttpStatus httpStatus) {
+    BoardCommandErrorCode(String message, HttpStatus httpStatus) {
         this.message = message;
         this.httpStatus = httpStatus;
     }
@@ -29,32 +32,32 @@ public enum BoardQueryErrorCode implements ErrorCode {
     }
 
     @Override
-    public BoardQueryException exception() {
-        return new BoardQueryException(this);
+    public BoardCommandException exception() {
+        return new BoardCommandException(this);
     }
 
     @Override
-    public BoardQueryException exception(Throwable cause) {
-        return new BoardQueryException(this, cause);
+    public BoardCommandException exception(Throwable cause) {
+        return new BoardCommandException(this, cause);
     }
 
     @Override
     public RuntimeException exception(Runnable runnable) {
-        return new BoardQueryException(this, runnable);
+        return new BoardCommandException(this, runnable);
     }
 
     @Override
     public RuntimeException exception(Runnable runnable, Throwable cause) {
-        return new BoardQueryException(this, runnable, cause);
+        return new BoardCommandException(this, runnable, cause);
     }
 
     @Override
     public RuntimeException exception(Supplier<Map<String, Object>> payload) {
-        return new BoardQueryException(this, payload);
+        return new BoardCommandException(this, payload);
     }
 
     @Override
     public RuntimeException exception(Supplier<Map<String, Object>> payload, Throwable cause) {
-        return new BoardQueryException(this, payload, cause);
+        return new BoardCommandException(this, payload, cause);
     }
 }
