+++ services/board/api/exception/src/main/java/nettee/board/BoardQueryErrorCode.java
@@ -1,20 +1,21 @@
-package me.nettee;
+package nettee.board;
 
 import java.util.Map;
 import java.util.function.Supplier;
+
+import nettee.common.ErrorCode;
 import org.springframework.http.HttpStatus;
 
-public enum BoardCommandErrorCode implements ErrorCode {
+public enum BoardQueryErrorCode implements ErrorCode {
     BOARD_NOT_FOUND("게시물을 찾을 수 없습니다.", HttpStatus.NOT_FOUND),
     BOARD_GONE("더 이상 존재하지 않는 게시물입니다.", HttpStatus.GONE),
     BOARD_FORBIDDEN("권한이 없습니다.", HttpStatus.FORBIDDEN),
-    DEFAULT("게시물 조작 오류", HttpStatus.INTERNAL_SERVER_ERROR),
-    BOARD_ALREADY_EXIST("게시물이 이미 존재합니다.", HttpStatus.CONFLICT);
+    DEFAULT("게시물 조작 오류", HttpStatus.INTERNAL_SERVER_ERROR);
 
     private final String message;
     private final HttpStatus httpStatus;
 
-    BoardCommandErrorCode(String message, HttpStatus httpStatus) {
+    BoardQueryErrorCode(String message, HttpStatus httpStatus) {
         this.message = message;
         this.httpStatus = httpStatus;
     }
@@ -30,32 +31,32 @@ public enum BoardCommandErrorCode implements ErrorCode {
     }
 
     @Override
-    public BoardCommandException exception() {
-        return new BoardCommandException(this);
+    public BoardQueryException exception() {
+        return new BoardQueryException(this);
     }
 
     @Override
-    public BoardCommandException exception(Throwable cause) {
-        return new BoardCommandException(this, cause);
+    public BoardQueryException exception(Throwable cause) {
+        return new BoardQueryException(this, cause);
     }
 
     @Override
     public RuntimeException exception(Runnable runnable) {
-        return new BoardCommandException(this, runnable);
+        return new BoardQueryException(this, runnable);
     }
 
     @Override
     public RuntimeException exception(Runnable runnable, Throwable cause) {
-        return new BoardCommandException(this, runnable, cause);
+        return new BoardQueryException(this, runnable, cause);
     }
 
     @Override
     public RuntimeException exception(Supplier<Map<String, Object>> payload) {
-        return new BoardCommandException(this, payload);
+        return new BoardQueryException(this, payload);
     }
 
     @Override
     public RuntimeException exception(Supplier<Map<String, Object>> payload, Throwable cause) {
-        return new BoardCommandException(this, payload, cause);
+        return new BoardQueryException(this, payload, cause);
     }
 }
