+++ services/board/driven/rdb/src/main/java/nettee/board/entity/type/BoardEntityStatus.java
@@ -1,12 +1,12 @@
 package nettee.board.entity.type;
 
-import me.nettee.type.BoardStatus;
+import nettee.board.type.BoardStatus;
 
 import java.util.Arrays;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import static me.nettee.BoardCommandErrorCode.DEFAULT;
+import static nettee.board.BoardCommandErrorCode.DEFAULT;
 
 public enum BoardEntityStatus {
     REMOVED(
