+++ services/board/driving/web-mvc/src/main/java/nettee/board/web/mapper/BoardDtoMapper.java
@@ -1,7 +1,7 @@
 package nettee.board.web.mapper;
 
-import me.nettee.Board;
-import me.nettee.type.BoardStatus;
+import nettee.board.Board;
+import nettee.board.type.BoardStatus;
 import nettee.board.web.dto.BoardCommandDto.BoardCreateCommand;
 import org.mapstruct.Mapper;
